+++ test/NLMPC/test_constraints.cpp
@@ -29,18 +29,22 @@ TEMPLATE_TEST_CASE_SIG(
     (2, 1, 1, 5, 5, 0))
 {
     constexpr int Teq = 0;
+    static constexpr auto sizer = mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq));
 
-    mpc::Constraints<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> conFunc;
-    conFunc.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Constraints<sizer>> conFunc;
+    conFunc = std::make_shared<mpc::Constraints<sizer>>();
+    conFunc->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    mpc::Mapping<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> mapping;
-    mapping.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Mapping<sizer>> mapping;
+    mapping = std::make_shared<mpc::Mapping<sizer>>();
+    mapping->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    mpc::Model<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> model;
-    model.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Model<sizer>> model;
+    model = std::make_shared<mpc::Model<sizer>>();
+    model->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    model.setContinuos(true);
-    model.setStateModel([](
+    model->setContinuous(true);
+    model->setStateModel([](
                             mpc::cvec<TVAR(Tnx)> &dx,
                             const mpc::cvec<TVAR(Tnx)> &x,
                             const mpc::cvec<TVAR(Tnu)> &u,
@@ -49,12 +53,12 @@ TEMPLATE_TEST_CASE_SIG(
             dx[0] = ((1.0 - (x[1] * x[1])) * x[0]) - x[1] + u[0];
             dx[1] = x[0]; });
 
-    conFunc.setModel(model, mapping);
+    conFunc->setModel(model, mapping);
 
     mpc::cvec<TVAR(Tnx)> x0;
     x0.resize(Tnx);
     x0 << 0, 0;
-    conFunc.setCurrentState(x0);
+    conFunc->setCurrentState(x0);
 
     // input decision variables vector
     mpc::cvec<TVAR(((Tph * Tnx) + (Tnu * Tch) + 1))> x;
@@ -68,7 +72,7 @@ TEMPLATE_TEST_CASE_SIG(
     costExpected.resize(Tph * Tnx);
     costExpected << 0, -1, -2, -2, -2, -2, -2, -2, -2, -2;
 
-    auto c = conFunc.evaluateStateModelEq(x, false);
+    auto c = conFunc->evaluateStateModelEq(x, false);
 
     REQUIRE(c.value == costExpected);
     REQUIRE(c.grad.isZero());
@@ -81,18 +85,22 @@ TEMPLATE_TEST_CASE_SIG(
     (2, 1, 1, 5, 5, 1))
 {
     constexpr int Teq = 0;
+    static constexpr auto sizer = mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq));
 
-    mpc::Constraints<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> conFunc;
-    conFunc.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Constraints<sizer>> conFunc;
+    conFunc = std::make_shared<mpc::Constraints<sizer>>();
+    conFunc->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    mpc::Mapping<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> mapping;
-    mapping.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Mapping<sizer>> mapping;
+    mapping = std::make_shared<mpc::Mapping<sizer>>();
+    mapping->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    mpc::Model<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> model;
-    model.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Model<sizer>> model;
+    model = std::make_shared<mpc::Model<sizer>>();
+    model->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    model.setContinuos(true);
-    model.setStateModel([](
+    model->setContinuous(true);
+    model->setStateModel([](
                             mpc::cvec<TVAR(Tnx)> &dx,
                             const mpc::cvec<TVAR(Tnx)>& x,
                             const mpc::cvec<TVAR(Tnu)>& u,
@@ -101,9 +109,9 @@ TEMPLATE_TEST_CASE_SIG(
         dx[0] = ((1.0 - (x[1] * x[1])) * x[0]) - x[1] + u[0];
         dx[1] = x[0]; });
 
-    conFunc.setModel(model, mapping);
+    conFunc->setModel(model, mapping);
 
-    conFunc.setIneqConstraints([](
+    conFunc->setIneqConstraints([](
                                    mpc::cvec<TVAR(Tineq)> &eq_con,
                                    const mpc::mat<TVAR(Tph + 1), TVAR(Tnx)>& x,
                                    const mpc::mat<TVAR(Tph + 1), TVAR(Tny)>&,
@@ -118,7 +126,7 @@ TEMPLATE_TEST_CASE_SIG(
     mpc::cvec<TVAR(Tnx)> x0;
     x0.resize(Tnx);
     x0 << 10, 0;
-    conFunc.setCurrentState(x0);
+    conFunc->setCurrentState(x0);
 
     mpc::cvec<TVAR(((Tph * Tnx) + (Tnu * Tch) + 1))> x;
     x.resize((Tph * Tnx) + (Tnu * Tch) + 1);
@@ -131,7 +139,7 @@ TEMPLATE_TEST_CASE_SIG(
     costExpected.resize(Tineq);
     costExpected << x0[0];
 
-    auto c = conFunc.evaluateIneq(x, false);
+    auto c = conFunc->evaluateIneq(x, false);
 
     REQUIRE(c.value == costExpected);
     REQUIRE(c.grad.isZero());
@@ -144,18 +152,22 @@ TEMPLATE_TEST_CASE_SIG(
     (2, 1, 1, 5, 5, 0))
 {
     constexpr int Teq = 1;
+    static constexpr auto sizer = mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq));
 
-    mpc::Constraints<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> conFunc;
-    conFunc.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Constraints<sizer>> conFunc;
+    conFunc = std::make_shared<mpc::Constraints<sizer>>();
+    conFunc->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    mpc::Mapping<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> mapping;
-    mapping.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Mapping<sizer>> mapping;
+    mapping = std::make_shared<mpc::Mapping<sizer>>();    
+    mapping->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    mpc::Model<mpc::MPCSize(TVAR(Tnx), TVAR(Tnu), TVAR(0), TVAR(Tny), TVAR(Tph), TVAR(Tch), TVAR(Tineq), TVAR(Teq))> model;
-    model.initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
+    std::shared_ptr<mpc::Model<sizer>> model;
+    model = std::make_shared<mpc::Model<sizer>>();    
+    model->initialize(Tnx, Tnu, 0, Tny, Tph, Tch, Tineq, Teq);
 
-    model.setContinuos(true);
-    model.setStateModel([](
+    model->setContinuous(true);
+    model->setStateModel([](
                             mpc::cvec<TVAR(Tnx)> &dx,
                             const mpc::cvec<TVAR(Tnx)>& x,
                             const mpc::cvec<TVAR(Tnu)>& u,
@@ -164,9 +176,9 @@ TEMPLATE_TEST_CASE_SIG(
         dx[0] = ((1.0 - (x[1] * x[1])) * x[0]) - x[1] + u[0];
         dx[1] = x[0]; });
 
-    conFunc.setModel(model, mapping);
+    conFunc->setModel(model, mapping);
 
-    conFunc.setEqConstraints([](
+    conFunc->setEqConstraints([](
                                  mpc::cvec<TVAR(Teq)> &eq_con,
                                  const mpc::mat<TVAR(Tph + 1), TVAR(Tnx)>& x,
                                  const mpc::mat<TVAR(Tph + 1), TVAR(Tnu)>&)
@@ -179,7 +191,7 @@ TEMPLATE_TEST_CASE_SIG(
     mpc::cvec<TVAR(Tnx)> x0;
     x0.resize(Tnx);
     x0 << 10, 0;
-    conFunc.setCurrentState(x0);
+    conFunc->setCurrentState(x0);
 
     mpc::cvec<TVAR(((Tph * Tnx) + (Tnu * Tch) + 1))> x;
     x.resize((Tph * Tnx) + (Tnu * Tch) + 1);
@@ -192,7 +204,7 @@ TEMPLATE_TEST_CASE_SIG(
     costExpected.resize(Teq);
     costExpected << x0[0];
 
-    auto c = conFunc.evaluateEq(x, false);
+    auto c = conFunc->evaluateEq(x, false);
 
     REQUIRE(c.value == costExpected);
     REQUIRE(c.grad.isZero());